Update patch set 5
Patch Set 5:

(1 comment)

I'm starting to wonder whether deduping makes any sense here at all. All the jobs are for batches of titles. These batches tend to be quite different. 

We could try to create a signature from the entity and aspects touched by a single Change from a ChangeNotificationJob, and use that as the root job sigature. Same aspect of the same entity should result in the same set of pages getting purged/updated.

Patch-set: 5