Update patch set 2
Patch Set 2:

Guys, I'm afraid this is not the place where this discussion should happen. We might discuss if we want to stick to a specific indention style, and if we want a rule that enforces this on all code, no matter how old it is. But this decision will not be made here in a Gerrit discussion thread.

As I already said: please disable the rule that is responsible for the +/- 2000 changes to unrelated lines of code in this patch.

Patch-set: 2