Update patch set 7
Patch Set 7:

The "root" job de-duplication is targeted towards the case of batches of titles, since it's based on the title of the template that was changed, not the specific batches of pages. The signature/timestamp combination is for consecutive edits to the same template. The use of timestamp alone is just custom logic in a few job types where it makes sense.

Patch-set: 7