Remove null special case from AliasesChangeOpDeserializer
As discussed in Idb23a44 and in person this is a new feature that is
not needed, not documented, and possibly unclear. If the user provides
null instead of a list of aliases, what does he mean? Remove? Was it
a mistake? If it was a mistake, why should the mistake be ignored and
not cause a proper error message? We think the error message makes more
sense.

Bug: T175009
Change-Id: Id9b7d315dd0cc74dbd51cf6088f68a0ee2f9f878