Update patch set 4
Patch Set 4:

RDF export is meant to export all values. This is meant to index only small subset of values relevant to searching (in fact, I'm not even sure it makes sense to keep many of them I've added, so far we only have use case for entyty ID). This is not an export format, and I don't want to make a 5-line function into a dozen of classes just for that. I need a small fixed list, that's the point.

Patch-set: 4