Update patch set 6
Patch Set 6: Code-Review+2

Thanks. Making a code public just to be able to test some in the end internal detail (it was meant as private, i.e. internal) is always a bit suspicious.
And in this case what is tested is not that WikibaseRepo would provide service generating correct URIs but rather that it passes the correct settings to the service. What I'd rather see here instead would be an edge-to-edge-style test testing that repo returns a service which does what it is expected, apart from testing individual components (which is done currently in "unit" tests, so that is OK already).
But this seems to me to be a part of a bigger topic, so I am fine with accepting this as it is for now.

Patch-set: 6
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2