Update patch set 2
Patch Set 2:

> Leszek, can you please explain the reasoning behind this? What motivated this patch? 

What the tool does is explained more than in the commit message on:
https://www.mediawiki.org/wiki/MinusX

The issue the tool fixes does not seem THAT critical but OTOH the cost of using the tool seems reasonable to me. It is of course open for discussion.

> How did you found this tool?

The author announced it on wikitech-l.

> Why do you think it's worth it?

As said above. Does not seem critical but does not hurt given the cost.

Patch-set: 2
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0