Update patch set 1
Patch Set 1:

that was log of each jenkins build says about it:

> You made a reference to a non-existent script @validate --no-interaction

So if running "composer validate" is intended as part of a "test" script, then the command should be fixed in my understanding. Fine with me, just does not make sense to have a broken command there.

AFAICT this seems also broken in several of "other code bases we have".
Or is it run but the warning is emitted to mislead me?

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0