Update patch set 1
Patch Set 1:

Note that both refactorings change the hash calculations. I believe this is fine. These job hashes are not persisted. There might be a mismatch happening for a short time, but the only effect it will have is a few missed opportunities for de-duplication.

A positive effect both refactorings might have is better de-duplication, because the page titles are ordered now. This was not guaranteed before.

Patch-set: 1
Label: Verified=0