Update patch set 1
Patch Set 1: Code-Review-1

Thanks for taking care of this. I'm willing to merge this simple fix. No problem with the code. But there must be a test covering this, to make sure the addition will survive future refactorings (and there will be refactorings, as one can tell from the TODOs).

Patch-set: 1
Label: Code-Review=-1