Fix statementview change propagation
The qualifiers and references are both a generic listview, so they can’t
be updated with a SnakList or a ReferenceList (the listview expects an
Array, or at least something subscriptable like in `value[ index ]`).
Instead, update the qualifiers with the group snak lists, and the
references with the array of references, mirroring their initialization
in _createQualifiersListView() / _createReferencesListview().

This is part of T176604, but doesn’t actually fix that bug yet: the snak
views are now properly reinitialized, but they don’t yet update their
hash on initialization. A separate commit (I7bce485a0d) will fix that.

Bug: T176604
Change-Id: I3455823bf332939569e3c1e5da13cb4414e100cf