Update hash when initializing snakview
For fresh snaks without underlying value, this just removes the
`wikibase-snakview-` class that is added by the template (which is
initialized with an empty string as hash) – see T176633. For snaks
initialized with a value from the database, this ensures that the snak
is also present in the element’s class list, fixing T176604.

Bug: T176604
Change-Id: I7bce485a0d76f4479b73449d586e3266e305b024