Index descriptions for Wikibase entites
This creates custom analyzers text, text_search, plain and plain_search
for every language, and assigns matching analyzer set to per-language fields.
Exception: text & text_search are not created for languages that are not
configured in useStemming option.

This is not ideal since many of those are the same, but we do not have access
to necessary data in DescriptionsField class to figure out which ones are the same.
Hopefully, having 1800 analyzers is not a problem...

Filters and char_filters are deduplicated so only necessary ones are there.

Bug: T176903
Change-Id: Ie304527c41c364e0bdb536cb04cb066fff6c6d7b