Update patch set 1
Patch Set 1: Code-Review+1

This is a very trivial baseline, there is nothing wrong with the code in here. All this patch basically does is copy pasting the existing label usage (L) code to be able to record description usage (D) separately.

My only issue with this patch is that all this code is put in places that feel awkward. I totally expected all this code to appear directly after the label usage code. I can imagine why you decided to put the code where it is now. However, can I ask you to move it next to the label-related code?

Patch-set: 1
Label: Code-Review=+1