Update patch set 2
Patch Set 2: Code-Review-1

Ok. I talked to Lydia and Daniel and we all think the new function should be called "getAllStatements" for a series of reasons. One is: when a user looks at the documentation, he should not get the impression that getStatements is the baseline he should use. We don't want this. The default should still be getBestStatements.

That's also why getAllStatements should be listed *below* getBestStatements.

Lydia and Daniel disagree with my idea to flick the parameter order. They think consistency with the existing function is more important than convenience. This is ok here in this patch. We might want to allow nil for the first parameter. But this can be done in a later patch.

I can amend this patch if you like. Please tell me.

Patch-set: 2
Label: Code-Review=-1