Update patch set 4
Patch Set 4:

I'm focusing on documentation and cleanup in this patch. Some things I do here might be a little controversial, because I'm partly undoing some of the changes done in the previous patch I02d8f7d (e.g. the introduction of Assert in EntityAccessos, and naming a variable "rank"). Marius, please tell me if you feel uncomfortable with one of these changes and I will minimize this patch to get it in a state that's fine for everybody.

Patch-set: 4
Label: Verified=0