Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

I checked and played around with this example:

https://de.wikipedia.org/wiki/Spezial:ApiSandbox#action=query&list=querypage&qppage=PagesWithBadges

The suggested change sounds good to me. An alternative would be to make the special page work with the "querypage" mode. But I can't see how that would work. And it's really not worth it because the exact same information is in the "page_props" table and can already be queries via other parameters of the API.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1