Remove unstubbing from Wikibase broadly
Unstubbing of objects is largely going away from MediaWiki core
in the near future. All of the instances of content language here
are already pointless as $wgContLang hasn't been stubbed in awhile.

Using RequestContext's instance of these objects is functionally
identical, without the complexity and overhead of attempting to
unstub a likely-already unstubbed object.

Change-Id: I0aa2442d4758a76c4e497f825f10743ea95677ec