Update patch set 1
Patch Set 1:

Never mind, I found the answer to my own question in the parent ticket: I724976d.

I'm not setting a +2 because I would love to have a second reviewer for the added tests. If nobody else feels like I will come back at this and merge it, probably on Friday.

Patch-set: 1