Split Echo hook handlers based on services needed
BeforeCreateEchoEvent runs when the Echo extension is
initialized, so on every page load.

All this hook needs is a couple settings and no services.
Other echo hooks require some services that are heavy to
instantiate.

This patch splits the hook handlers into two classes,
depending on what services/dependencies they need so we
avoid instantiating unnecessary services on page load.

Bug: T177311
Change-Id: I3c13b6d3a88a9db706db9b6a7c753080994b8830