Update patch set 2
Patch Set 2: Code-Review+1

(3 comments)

Full review done. The code looks good for me. The only thing I'm not happy with is the amount of bool parameters we are passing around. That starts to be seriously confusing. I will upload a follow up and try to improve this situation.

Patch-set: 2
Label: Code-Review=+1