Refactor PropertyRdfBuilder logic for readability
The main reason to touch this code was the one method having 3 bool
parameters in a row. Note that these 3 parameters are not independent,
and it was possible to pass combinations of bools that do not make that
much sense.

I'm using named string constants now, and reduced the 3 bool parameters
to 2 strings.

I also extracted some code snippets just so I can label them with a
descriptive variable name.

Bug: T121274
Change-Id: I36f381cf3f7397efc8c9e902275c0cf91a0a6a07