Update patch set 20
Patch Set 20: Code-Review+1

(14 comments)

Full review done again. I see quite a few missing and not-that-optimal snippets of code in this patch. But I believe we can merge it as it is, try it out, and continue working on the code later. I just want to make sure we do not forget about these possible issues and improvements. I will upload a follow up patch with my suggestions and link it to the same ticket to make my suggestions visible, without blocking this patch from being merged.

Patch-set: 20