Add wikibase-snakview-indicators span to snaks
This alters the DOM structure of snaks to wrap the
wikibase-snakview-value div inside a new wikibase-snakview-body div,
together with a new wikibase-snakview-indicators div. The two divs
inside the body are made to appear next to each other (unlike normal
block elements) via CSS.

The new wikibase-snakview-indicators element added with this change is a
container where any external code can append elements belonging to the
snak without any unwanted interation between those elements and
Wikibase’s own elements. The indicators are hidden while editing a
statement, and are cleared when the statement is saved; the external
code can subscribe to the wikibase.statement.saved hook to re-populate
them afterwards.

Bug: T95403
Change-Id: I1399233617581eb3b600fb23ea1de1058ebe4811