Quick fix for flooding of the RC table.
Under some conditions, Wikibase can trigger a very large number (millions)
of recentchanges rows to be inserted as the result of a single edit. These
inserts are batched according to the recentChangesBatchSize setting.

This patch aborts the injection after the first batch, to mittigate
the problem if flooding the rc table. This is not a permanent solution,
please refer to T177707 for discussion. ONe possible refinement would be
to inject rc records only for the pages with the most watchers.

Bug: T177707
Change-Id: I0162cbfd6db3ef147745d281a7568313e998f0c3