Update patch set 10
Patch Set 10:

(Haven't looked at the tests yet) The code looks good.

But for this to be actually useful right now (we still need X or O usages along statement usages), we need more fine-grained configuration. I suggest turning fineGrainedLuaTracking into a bool|string[] which either enables/ disables the fine grained tracking for everything (bool) or enables the fine grained tracking for just certain aspects (fineGrainedLuaTracking would be ['L', 'C', …] in that case). Also, we need a separate configuration controlling whether to record an X usage. If we have that, we can enable this selectively on certain wikis and see how it holds up.

Patch-set: 10