Run QUnit tests with karma runner
This all is still far from perfect. Not all code is tested,
tests are testing implementation details, not behaviour, etc.

This change is not intending to improve this situation, it
just allows to have current tests run with needing to
use MediaWiki QUnit test runner.

RepoApi tests are simplified, so there no need to depend
on the actual mw.Api. A very simple mock of mw object
is provided to allow running RepoApiError tests.

Bug: T178088
Change-Id: I37c10524f181b7bab4d40b5a02a869ab7f4d96aa