Update patch set 1
Patch Set 1:

> (1 comment)

Yeah, that's one of the reasons this is WIP; it can't be tested properly until the new version of purtle is actually released (with whatever version # is chosen).

The other reason is https://gerrit.wikimedia.org/r/#/c/384050/1/repo/tests/phpunit/includes/LinkedData/EntityDataSerializationServiceTest.php
which needs some updating w/ the proper regexp for the JSON-LD output.

Patch-set: 1