Update patch set 3
Patch Set 3: Code-Review+1

(2 comments)

I can understand I reasoning behind the change provided in the ticket.
The change actually also affects (or might affect) services which or not "dump-related", but this still seems OK.

As I am not familiar with all details and consequences of whether the cache should be used, written in a particular context, and of all of these, I'd wait for someone more familiar with this to have a look and verify.

Patch-set: 3
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1