Update patch set 3
Patch Set 3: Code-Review-2

(3 comments)

This is super-awkward, and super-scary. I'm setting a temporary -2 so this does not get merged accidentally. This is not meant to block this patch forever, but to have the needed discussion first.

The main issue I have is the (basically) boolean flag that shortcuts half of the CachingEntityRevisionLookup, and makes it behave super-unpredictable. It will never cache anything, except somebody calls entityUpdated() explicitly. If this is not called, then getEntityRevision() will fetch everything live from the database.

This appears to me like this way of caching only certain things is implemented on the wrong level. I find it obvious it does not belong to this class, as it destroys it's contracts entirely. I assume what is needed here is an other implementation of an EntityRevisionLookup that can be pre-filled with certain information.

Patch-set: 3