Update patch set 11
Patch Set 11: Code-Review-1

(6 comments)

Lua integration tests for this would also be nice… although I see that this is not trivial to set up. Probably warrants an entirely new test case (a little similar to Scribunto_LuaWikibaseLibraryNoLinkedEntityTest or Scribunto_LuaWikibaseLibraryInProcessEntityCacheTest).

We should only merge this together with a patch introducing T178153.

Patch-set: 11