Make Item… and PropertyFieldDefinitions accept arrays
This is the outcome of a discussion with Ladsgroup when we reviewed
Ib52a5fd together. Basically, the three separately named constructor
parameters don't really have a benefit. They can be confused, and not
even something bad would happen. The order just does not matter (except
two are returning array elements with the same keys, which is not
relevant in production right now).

Bug: T175199
Change-Id: I74703ea409d5102db9c10ea0ac09cae7fb75b994