Update patch set 3
Patch Set 3: Code-Review-1

(3 comments)

"This is only […] about reading from the cache, but avoiding to write to it". I believe I kind of understand what the intention here is. But try to explain this sentence to somebody who is new to our code base. He will ask what "a cache you can not write to" is for.

I hope we can find a better place in the code to implement the specific caching layer you want. See my inline comments. It will probably need a different implementation there.

Patch-set: 3
Label: Code-Review=-1