Add missing @license and initial @author tags
I found these candidates by searching for classes that start with a
comment, but the comment does not contain a @license tag. I added the
initial @author that introduced the class, but I did this only if the
same author name already appears in other classes in the same code base.
Some people never added their name as an @author, so I left them out.

Change-Id: I1faa844f9695f6e5e78df6ce23b993f0659aa754