Update patch set 1
Patch Set 1:

Do we need to do this? We already have license terms in COPYING in the main repo, and authors are tracked in git, also in cooperative environment it is not that useful to have author in each file - the code may be maintained by somebody else now, and probably more than one person worked on it.

Patch-set: 1