Update patch set 1
Patch Set 1:

Code is not only visited by people using git. It shows up in other places, sometimes even copied and disconnected from where it came from.

Repeating the @license tag is the bare minimum we clearly agreed on, as you can see, because all files except a few already do this.

Listing at least the one @author who started (and mainly wrote) a file is also good practice. It helps experienced developers to understand much faster how code came to live, why it is structured and named as it is, and so on. And for new users it gives a nice personal touch. They can literally go and ask the person, or start a conversation with somebody else if the original developer is not around any more.

There are many more reasons why this is good practice.

Honestly, just looking at the existing code base should be proof enough.

Patch-set: 1