Do not use messages from View-specific dependency in HtmlTimeFormatter
valueview-expert-timevalue-calendar-gregorian/julian messages
are defined in data value view values JS library, which is
a dependency of the View component.
In principle, these are not known to the Lib component, as
it is meant to be independent from View component.
Classes in the Lib component should hence only use messages
which are known inside of this component.

Change-Id: I4c893a1ee8f7c9bbe2c962636a16f865757f05f9