Use default values when empty "props" provided to wbgetentities API
This change is made in the assumption that is actually what is
expected to happen when providing props= in the API call.

Background:
recent change in the Wikibase JS API library
4f5770baf9f256e6530ede7345314e78917b12c8
(a.k.a. I84d360f7532c28353aff39baa791cbeffe7b6768)
caused Wikibase UI widget call wbgetentities with the empty
props parameter when e.g. editing the statement with the value
being entity ID (previously props parameter used to be skipped
from the API call entirely). With the empty props list, API
respons has been really sparse (has only contained the ID
and entity type), and the front-end code has been trying
to get non-existing data from the API response, leading to
bugs in the UI.

While the issue is on the side of the front-end code,
it might actually be stemming from the unexpected
behaviour of the wbgetentities API.

Change-Id: If33d13243007203afe117a11f47b469a25595844