Update patch set 1
Patch Set 1: Code-Review+2

(2 comments)

I will merge this for the moment to unblock us and be able to move forward. But I strongly suggest to not make the same mistake again: please do not implement features like this via subclassing! Whenever we did this in the past it turned out to be a big mistake.

Can we talk about possible strategies tomorrow, when I'm back in the office?

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2