Update patch set 4
Patch Set 4:

WikibasePrefixSearcher is implicitly covered by EntitySearchElastic tests, since it uses WikibasePrefixSearcher to deliver actual search. It's kinda hard to test them individually as they are both needed to work properly. Eventually, in the shining distant future, I'd like to get rid of WikibasePrefixSearcher, but for that Searcher needs to be much more flexible and modularized than now. It's happening, but slowly. We could add @covers WikibasePrefixSearcher to the EntitySearchElastic tests, probably.

Patch-set: 4