Update patch set 2
Patch Set 2:

(3 comments)

I'm super fine with the idea of this service, but I feel the names of both the interface and it's only method are not optimal. For a start, why is it called a "by title" lookup, but the getter asks for some kind of "link"?

So what this interface realize: the reverse mapping back to an entity ID, based on any global wiki page title in the cluster:
* Can be a file description page and it's associated MediaInfo entity.
* Can be a Wikipedia article and it's associated Wikidata item.
* Could even be a page in the item namespace and it's item ID.

The only critical requirement is that these mappings must be strict 1:1, which they all are at the moment.

* EntityIdByGlobalTitleLookup
* EntityIdByLinkTargetLookup

Method names can be:
* resolveEntityId
* lookupEntityId
* getLinkedEntityId

I suggest to not repeat any "…by…" wording in the method name, because this should be super-obvious from both the parameters as well as the variable name of the service itself. For example:

$entityId = $entityIdByLinkTargetLookup->getLinkedEntityId( 'enwiki', 'Berlin' );
$entityId = $entityIdByLinkTargetLookup->getLinkedEntityId( 'commons', 'File:X.jpg' );

Patch-set: 2