Update patch set 1
Patch Set 1:

Oh. Oh! I hope this is not something we are actually going to do any time soon. The biggest possible issue I believe to see (not 100% sure I'm right about this) is that the hook is added to one of possibly many specific implementations, but the code that uses this implementation intentionally binds against an interface.

I think we need to look at the structure that is already laid out in WikibaseMediaInfo.entitytypes.php and it's parents in the Wikibase.git code base and either change one of the services in this file the way you need it to change, or introduce a new one.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0