Update patch set 5
Patch Set 5:

(2 comments)

I'm not sure whether this is actually meant to run the search, but it looks like it needs to set up a number of profile variables if it does. Right now this code won't find any search profiles, so it won't execute any searches. I may not be fully understanding this yet.

Patch-set: 5
Label: Verified=0