Update patch set 2
Patch Set 2:

(1 comment)

If we no longer use lockGraceIntervale, the --lock-grace-interval option and all handling for it should be removed.

Note: lockGraceIntervale is intended to work around "lost" locks. If a client wiki has been marked as locked for a long time, we should should try to acquire the lock anyway, ignoring the state in the table.

Patch-set: 2