Check wbEntityId instead of wbEntity in entityViewInit
This is just a safety check anyways, in case someone loads the module on
non-entity pages. Even if there is a wbEntityId in the config but no
wbEntity data available, there shouldn’t be a problem – the
wikipage.content hook handler will just set up some promises that are
never resolved (because wikibase.entityPage.entityLoaded never fires)
and event handlers that are never triggered.

Bug: T179699
Change-Id: Id5b73b79d7795cbf2865750ac58d312329fe31b0