Update patch set 4
Patch Set 4:

(5 comments)

Is the fact this needs approval by PM reason enough to set a -1? I'm not sure. I mean, we can merge this  as it is and add backwards compatible code paths later. We just need to make sure we don't forget about this before this new code is finally deployed to production.

So no -1 for now, but please make sure the concerns I expressed are considered.

Patch-set: 4
Label: Verified=0