Update patch set 1
Patch Set 1:

(3 comments)

I understand that the linked ticket "is about an approach where […] FormIds are not EntityIds". When I look at what needs to be changed to allow for such an approach (and I assume the 3 places marked with a TODO here are not all that need to be changed), I find this approach scary.

To be honest I find the other approach also scary. :-(

Patch-set: 1