Fix minor type violation in LexemeHandler "history" action override
We know this does not make a difference because there are only two
classes implementing the Page interface. But the code does not know.
I get a warning that Article::newFromWikiPage can only be called with
a WikiPage, which is true and what the if() above should check for.

See I55edb34 as well as Icf6055f for reference.

Change-Id: Ia401c27318eee8adcf938884f5ec038d96f5fc86