Update patch set 1
Patch Set 1: Code-Review-1

> What did we miss?

It might be that this is indeed all that needs to be changed to make wbsetclaim work with statements on Forms.

… until we talk about statements on Senses.

… until we talk about the next API module.

… and the next.

One of the TODOs in this patch is about GUIDs, which is something that is specific for statements, and can probably be reused for all other API modules that do something with statements.

The two other TODOs are about a ChangeOp factory, and the same for summary lines. Both services are specific for the operation in question, which means we will end with a super-complex entity type definition map that contains a large set of ChangeOp factories as well as summary builders for each and every API module. And a lot of this code will be duplicated between different code bases, but these duplications can not be independent from each other but must strictly follow a common specification. Where does this specification live?

TL;DR: That's a lot of code in a lot of places.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1