To identify superseded requests, use the requested search term instead of the returned search term
Previously, the search suggestions dropdown would be updated only if the
the search term in the response matched the most recently dispatched
search term. This was supposed to suppress out-of-order responses. However,
it had the undesired effect of suppressing all responses in which the
search term is changed by MediaWiki's Unicode normalisation.

It's not sufficient to run NFC on the client-side, since MediaWiki's
Unicode cleanup routine is not precisely the same as NFC. Also, NFC
changes from one Unicode version to the next. And also, it is slow.

So, use the term precisely as requested to identify responses. It is
already available in the relevant closure.

Bug: T170779
Change-Id: Id0d1d081e921a0f24a380260dcb5186d98e443a9