Update patch set 2
Patch Set 2:

Regarding the change itself: I would not say I like the coupling this introduces, but as long as it is in private method I could live with that.

Regarding the test failure: there is clearly an issue/confusion about what does $diff->isEmpty() and empty( $diff ) mean, and how do they differ. I've made a change as seen in PS2 which made the test failing on PS1 (the failure was actually correct) pass for me. I guess to match exactly the old code it should be: if ( !$labelsDiff->isEmpty() && !empty( $labelsDiff ) ), although maybe just checking isEmpty is enough. As I am not sure about the details, I hope someone could shed some light on this.

Patch-set: 2