Update patch set 2
Patch Set 2: Code-Review+1

Diff::isEmpty is a surprisingly expensive count internally. That's why I wanted to avoid it by utilizing a direct count() call on the ArrayObject, which was a mistake. I think I understand the difference now. Thanks for the fix.

Regarding "coupling": Either you are using the word in a sense that is different from what it means, or you have a very different understanding of code quality. Taking all the issues into account that piled up the past weeks and months, I believe it is the later. We seriously need to talk and find out what the underlying issue is that causes all the related friction recently.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1