Update patch set 2
Patch Set 2:

I believe my concerns still hold. The names that make up the new interface introduced here are not optimal, and can be improved very easily. I'm willing to do this if somebody could suggest a way forward, and if it's fine if I change this patch.

It might also be that this patch is not needed any more and can be abandoned.

Patch-set: 2
Label: Verified=0