Update patch set 1
Patch Set 1: Code-Review+1

This is not currently supported on any locking backend, so technically it's fine to ditch this.
Given we don't plan to re-introduce this right now, I can see that we want to remove it for now.

Patch-set: 1
Reviewer: Hoo man <50@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1