Update patch set 1
Patch Set 1:

honestly I am not sure if it works or not.
Locally run on its own, the test passes for me, with the incomplete Wikibase autoload file.

I've tried to test it in I3eb74b65608ef899e3b3e6ba5f8acf87853c15df but it seems that either the said test in core passes, or it is not even run but the execution crashes when running some of Wikibase tests using the undeclared class.

So basically: I don't know :)

Patch-set: 1