Allow EntityHandlerTest to be used with Lexemes, etc.
This is needed to allow EntityHandlerTest to be used as a base class
for the respective tests for Lexeme and MediaInfo. Which in turn
is needed to make sure we are consistently asserting the contract
of EntityHandler.

Usinga base class for this isn't great, a trait would be better. That
change is left for later.

Change-Id: I487a4fcc62aa2ad9a4aeebab4d871583bbe876b8