Adding .luachecrc with a minimum configuration
The linter "luacheck" need a special configuration file to make it stop
complaining. In addition to the configuration file a few changes are
made to remove specific lints, thus removing the the need for inline
options and file-specific handling.

This does not add the linter to the toolchain, it is only for
in-editor use.

Bug: T180925
Change-Id: I091b46fdf358c36f02b36a65223f643a2c9e41b2