Update patch set 1
Patch Set 1:

> I'm super fine with the underscores, but everything else … I don't
 > know. I feel the comments clutter the code more than they help. Why
 > do they mention a specific tool? What is the benefit if this is in
 > nobodys toolchain?

They are inline directives to luacheck to turn off specific warnings
at the line where they are posted. For the few places where these are
posted it is also possible to use the Lua common solution of adding an
underscore. This works everywhere where a warning is triggered by a
variable.

Note that luacheck is one of the common tools to do linting in Lua,
and I am a bit surprised that it is not added already. I use this
linter, usually through grunt and an extension in vsc (vscode-luacheck),
but as there are no official grunt-plugin I only provide these fixes.

It is pretty straight forward to make a grunt-plugin from the existing
one for luahint. Note that luahint is a lot simpler than luacheck, but
the plugin would be pretty much the same.

Patch-set: 1