Also autoload client class in repo outside of tests
TitleFactory is used in repo’s own code, not just in the tests, so we
should still make sure the client class is loaded even outside the
tests. This is necessary on repo wikis with client functionality
disabled. (This fixes Icb197448fc, which in turn fixes Ic3c3f2b969.)

Change-Id: Ie5f7e6195facd4b7bc6521f48fc9c756474f8002