Fix flash of unstyled content (FOUC) on sitelinkgroupview
To see this issue, go to an item with multiple sitelinks (e.g.
https://www.wikidata.org/wiki/Q9410185 ), and watch the page loading.
You will see the word "[Collapse]" appear for a second before it is
replaced with an arrow.

This is because the .css file that replaces the word with the arrow
is loaded to late, via JS instead of in the head.

This patch moves the relevant CSS to a .css file that is loaded in
the head. I argue this is a good place for this CSS because the
relevant "mw-collapsible" is set on this specific
"sitelinkgroupview" template.

Bug: T158059
Change-Id: I7ab015ae1e2bc6a7401c37620c7eebea23b055a3