Make .valueview inline as well
This is a follow-up to commit 5c3e109dc2 (change I1399233617), which
makes the wikibase-snakview-value div inline when not in edit mode, so
that it can appear next to the indicators instead of having them both in
separate lines. When the value is created by the client-side JS code
instead of the PHP version, it is wrapped in another valueview div,
which also needs to be made inline to achieve the desired effect.

Bug: T95403
Change-Id: I3d35aa0deaa9cc1f7d2427e6ace964eb5cc51ceb