Update patch set 2
Patch Set 2:

Note: I first tried making all children (CSS selector *) of both the value and the indicators inherit the display, but that made the icons higher than they should be because some OOui-internal spans now inherited the wrong display value. We could still consider applying this to all valueview children of both value and indicators (reusing the selector in L235-236 instead of copying the value half), or applying this to all children of the value (using * instead of .valuelview), but for now I went with the most specific selector.

Patch-set: 2