Update patch set 36
Patch Set 36:

thinking about it I wonder if we should not move the ConfigBuilder code in cirrus, it has hard dependencies to cirrus and no strong deps to wikibase. As described in the tests there's a fear that we end up with "hard to change" codes requiring forcing +2 when the tests fail.
The hook hack could (perhaps) be fixed more easily?

Patch-set: 36