Update patch set 36
Patch Set 36:

I'm not against moving it - significant part of it is used only by Wikibase, but we could of course move it to Cirrus and expose it to Wikibase as a client, at least ConfigBuilder part. I don't think it would solve the hook problem in any case, since the hook application is Wikibase-specific. We would also need to think what to do if two extensions try to use this functionality. 
I'd prefer to merge this as is (provided there's no other problems of course) and then I'll look into moving the bulk of the code to Cirrus. The reason for that is reindexing Wikidata takes a lot of time, so I would rather work on refactoring while index is running than wait for refactoring being ready and block on that.

Patch-set: 36
Reviewer: Smalyshev <2096@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0