Update patch set 36
Patch Set 36: Code-Review+2

sure. As for the hook problem I was thinking that the code being in cirrus it'd be easier to introduce a new hook where extensions could declare the list of additional languages they'd want to have in the analysis config, the hook would be called prior to build the config.

Patch-set: 36
Reviewer: DCausse <2638@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2