Avoid using EntityDiffChangedAspectsFactory like a static constructor
Let's please be honest and use an actual static constructor method when
code really is called like a static constructor method.

I understand that ( new ClassName() )->methodName() is not meant to stay
like this. Which is why I'm changing it right now, before it becomes
technical debt we need to track and come back to.

We can change it back to use the factory (when it starts behaving like
an actual factory) any time. The fact that I'm introducing two static
constructor methods does not block us from doing so in the future.

Bug: T113468
Change-Id: Id17259c2c62f60a73a01d2243e72d0d6203dff6e