Update patch set 1
Patch Set 1:

@Leszek *calling* code shouldn't need to know about decorators... In my mind, constructing and wiring the services is application wiring, so it goes into WikibaseRepo. The actual decorator goes into WikibaseServices if it is sufficiently generic, or into a specialized component if it's specialized. 

If the concept of EntityContainers and HierarchicalIds is part of the PHPdata model, generic handling of these should be in WikibaseServices. IF otoh we'D implement this specifically for Forms and Descriptions, the decorator would live in the Lexeme extension; but then we'd need a way for the Lexeme extension to control the wiring in WikibaseRepo.

Patch-set: 1
Label: Code-Review=0