Update patch set 1
Patch Set 1:

> Does the code calling decorators better fit WikibaseRepo or WikibaseServices component?

I'm not sure I understand the question. My idea of a decorator is that there is only one call in Repo, but it doesn't know it is calling a decorator.

Sure, moving classes and interfaces to WikibaseServices is always welcome, if they fit there. But this is pretty much independent from what this patch does.

It might be that I'm using the word "decorator" in an odd way, and there is a better word for the pattern I'm using here. Please suggest a better term.

What Daniel writes is exactly what I had in mind.

Patch-set: 1