Catch uncatched SerializationException in EntityContent::getNativeData
This is quite obviously a bug. Whenever a serializer is called, code must
expect the serialization to fail, and handle the failure appropriately.
However, this was never necessary here because it was impossible to feed an
"invalid" entity to this serializer. The concept of an entity that can not
be serialized never existed, but exists now with StorageLexemeSerializer.

Personally I think StorageLexemeSerializer violates a contract and must be
changed. But this is independend from this patch. This try-catch should be
here no matter what.

I decided to simply return the entity object as a fallback. Could also
return an empty array, "really anything", as the interface says.

Bug: T182082
Change-Id: I7a2fec4f0ed6b1809806f6a56b64fc1c78c17abe